+++ db/migrate/002_rename_schedules_to_tasks.rb
@@ -1,0 +1,5 @@
+class RenameSchedulesToTasks < ActiveRecord::Migration
+  def change
+    rename_table :weekly_schedules, :recurring_tasks
+  end
+end