+++ lib/redmine_recurring_tasks/hooks/issue_view_hook.rb
@@ -2,7 +2,7 @@ module RedmineRecurringTasks
   module Hooks
     # Issue sidebar hook for show weekly schedule on issue page
     class IssueViewHook < Redmine::Hook::ViewListener
-      render_on :view_issues_show_description_bottom, partial: 'weekly_schedules/issues/bottom'
+      render_on :view_issues_show_description_bottom, partial: 'recurring_tasks/issues/bottom'
     end
   end
 end