+++ CHANGELOG.md
@@ -1,5 +1,8 @@
-# 0.0.3 (In progress)
+# 0.1.0
 
+* Add logging errors via logger
+* Add ability to choose months and month days
+* Change model from WeeklySchedule to RecurringTask
 * Fix schedule checking method
 * Move schedule to issue details bottom